cacheable - adding in hash store for memory
